+++ ricaun.Revit.UI.Tests/Items/Items/RevitComboBoxTests.cs
@@ -31,6 +31,74 @@ namespace ricaun.Revit.UI.Tests.Items.Items
             Assert.IsNotNull(comboBox.Image);
         }
 
+        [Test]
+        public void ComboBoxMember_SetImage_ShouldBe_Image()
+        {
+            var data = ribbonPanel.NewComboBoxMemberData();
+            Assert.IsNull(data.Image);
+            data.SetImage(BaseImage.Revit);
+            Assert.IsNotNull(data.Image);
+        }
+
+        [Test]
+        public void ComboBoxMember_SetLargeImage_ShouldBe_Image()
+        {
+            var data = ribbonPanel.NewComboBoxMemberData();
+            Assert.IsNull(data.Image);
+            data.SetLargeImage(BaseImage.Revit);
+            Assert.IsNotNull(data.Image);
+        }
+
+        [TestCase("Text")]
+        [TestCase("Name")]
+        public void ComboBoxMember_Create_Name_ShouldBe_ItemText(string name)
+        {
+            var comboBoxMember = comboBox.CreateComboBoxMember(name);
+            Assert.AreEqual(name, comboBoxMember.ItemText);
+        }
+
+        [TestCase("Text")]
+        [TestCase("Name")]
+        public void ComboBoxMember_Create_SetTest_ShouldBe_ItemText(string text)
+        {
+            var comboBoxMember = comboBox.CreateComboBoxMember();
+            comboBoxMember.SetText(text);
+            Assert.AreEqual(text, comboBoxMember.ItemText);
+        }
+
+        [TestCase("Group")]
+        [TestCase("Name")]
+        public void ComboBoxMember_Create_GroupName_ShouldBe_GroupName(string groupName)
+        {
+            var comboBoxMember = comboBox.CreateComboBoxMember(groupName: groupName);
+            Assert.AreEqual(groupName, comboBoxMember.GroupName);
+        }
+
+        [Test]
+        public void ComboBoxMember_Create_SetLargeImage_ShouldBe_Image()
+        {
+            var comboBoxMember = comboBox.CreateComboBoxMember();
+            Assert.IsNull(comboBoxMember.Image);
+            comboBoxMember.SetLargeImage(BaseImage.Revit);
+            Assert.IsNotNull(comboBoxMember.Image);
+        }
+
+        [Test]
+        public void ComboBoxMember_Create_SetImage_ShouldBe_Image()
+        {
+            var comboBoxMember = comboBox.CreateComboBoxMember();
+            Assert.IsNull(comboBoxMember.Image);
+            comboBoxMember.SetImage(BaseImage.Revit);
+            Assert.IsNotNull(comboBoxMember.Image);
+        }
+
+        [Test]
+        public void ComboBoxMember_Create_ShouldBe_Current()
+        {
+            var comboBoxMember = comboBox.CreateComboBoxMember();
+            Assert.AreEqual(comboBoxMember, comboBox.Current);
+        }
+
         [TestCase(1)]
         [TestCase(2)]
         [TestCase(3)]